+++ etna/metrics/__init__.py
@@ -7,6 +7,7 @@ from sklearn.metrics import r2_score
 from etna.metrics.base import Metric
 from etna.metrics.base import MetricAggregationMode
 from etna.metrics.functional_metrics import mape
+from etna.metrics.functional_metrics import max_deviation
 from etna.metrics.functional_metrics import rmse
 from etna.metrics.functional_metrics import sign
 from etna.metrics.functional_metrics import smape
@@ -19,6 +20,7 @@ from etna.metrics.metrics import MSLE
 from etna.metrics.metrics import R2
 from etna.metrics.metrics import RMSE
 from etna.metrics.metrics import SMAPE
+from etna.metrics.metrics import MaxDeviation
 from etna.metrics.metrics import MedAE
 from etna.metrics.metrics import Sign
 from etna.metrics.utils import compute_metrics